fix: move registered oracles context files closer together
