+++ packages/apps/human-app/frontend/src/modules/worker/views/registration/registration.page.tsx
@@ -13,7 +13,7 @@ import {
 } from '@/modules/worker/services/registration-in-exchange-oracles';
 import { useGetOracles } from '@/modules/worker/services/oracles';
 import { routerPaths } from '@/router/router-paths';
-import { useRegisteredOracles } from '@/shared/hooks/use-registered-oracles';
+import { useRegisteredOracles } from '@/shared/contexts/registered-oracles';
 
 export function RegistrationPage() {
   const navigate = useNavigate();
