+++ packages/apps/human-app/frontend/src/shared/contexts/registered-oracles/index.ts
@@ -1,0 +1,2 @@
+export * from './registered-oracles';
+export * from './use-registered-oracles';
