+++ packages/apps/human-app/frontend/src/shared/contexts/registered-oracles/use-registered-oracles.tsx
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { RegisteredOraclesContext } from '../contexts/registered-oracles';
+import { RegisteredOraclesContext } from './registered-oracles';
 
 export const useRegisteredOracles = () => {
   const context = useContext(RegisteredOraclesContext);
